requirements.txt
fastapi~=0.115.9
uvicorn~=0.34.2
langchain~=0.3.25
openai~=1.76.2
python-dotenv~=1.1.0
chromadb~=1.0.7
sentence-transformers~=4.1.0
google-generativeai
pypdf
pydantic~=2.11.4
protobuf~=5.29.4
numpy~=2.2.5
pillow~=11.2.1
h11~=0.16.0
pip~=25.1.1
attrs~=25.3.0
distro~=1.9.0
filelock~=3.18.0
rsa~=4.9.1
pyasn1~=0.6.1
grpcio~=1.71.0
rich~=14.0.0
Pygments~=2.19.1
markdown-it-py~=3.0.0
requests~=2.32.3
PyYAML~=6.0.2
zipp~=3.21.0
typing_extensions~=4.13.2
sniffio~=1.3.1
importlib_metadata~=8.6.1
pyproject_hooks~=1.2.0
build~=1.2.2.post1
click~=8.1.8
httpcore~=1.0.9
idna~=3.10
httpx~=0.28.1
zstandard~=0.23.0
jiter~=0.9.0
mdurl~=0.1.2
charset-normalizer~=3.4.2
setuptools~=80.1.0
overrides~=7.7.0
proto-plus~=1.26.1
regex~=2024.11.6
scipy~=1.15.2
threadpoolctl~=3.6.0
torch~=2.7.0
mpmath~=1.3.0
sympy~=1.14.0
onnxruntime~=1.21.1
tqdm~=4.67.1
triton~=3.3.0
packaging~=24.2
networkx~=3.4.2
urllib3~=2.4.0
fsspec~=2025.3.2
typer~=0.15.3
shellingham~=1.5.4
aiohttp~=3.11.18
yarl~=1.20.0
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.4.26
google-api-python-client~=2.169.0
grpcio-status~=1.71.0
httplib2~=0.22.0
MarkupSafe~=3.0.2
joblib~=1.4.2
websockets~=15.0.1
anyio~=4.9.0
pydantic_core~=2.33.2
orjson~=3.10.18
PyPika~=0.48.9
uvloop~=0.21.0
asgiref~=3.8.1
backoff~=2.2.1
starlette~=0.45.3
posthog~=4.0.1
langchain-core~=0.3.58
six~=1.17.0
python-dateutil~=2.9.0.post0
scikit-learn~=1.6.1
httptools~=0.6.4
watchfiles~=1.0.5
importlib_resources~=6.5.2
tenacity~=9.1.2
bcrypt~=4.3.0
jsonschema~=4.23.0
kubernetes~=32.0.1
mmh3~=5.1.0
chroma-hnswlib~=0.7.6
greenlet~=3.2.1
pyparsing~=3.2.3
oauthlib~=3.2.2
Deprecated~=1.2.18
annotated-types~=0.7.0
langchain-community~=0.3.23
langsmith~=0.3.41
langchain-text-splitters~=0.3.8
SQLAlchemy~=2.0.40
requests-toolbelt~=1.0.0
Jinja2~=3.1.6
websocket-client~=1.8.0
wrapt~=1.17.2
referencing~=0.36.2
jsonpointer~=3.0.0
jsonschema-specifications~=2025.4.1
durationpy~=0.9
requests-oauthlib~=2.0.0
tokenizers~=0.21.1
humanfriendly~=10.0
coloredlogs~=15.0.1
flatbuffers~=25.2.10
transformers~=4.51.3
safetensors~=0.5.3
uritemplate~=4.1.1
huggingface-hub~=0.30.2
jsonpatch~=1.33
google-auth-httplib2~=0.2.0